Completely ts-morph approach
This shows how we could create a pipeline of ts-morph based transforms, parse -> tsmorph xN -> serialize. It doesn't use the jscodeshift API at all, and avoids multiple parse/serialize steps.

Project is still messy (e.g. no jsdoc, bad filenames, old files left in place) but this shows the idea.
