+++ lib/rewriter.js
@@ -1,26 +1,39 @@
 'use strict';
 
-const cleanConsole = require('./mods/clean-console.js');
-const typer = require('./mods/typer.js');
+const assert = require('assert');
+const ts = require('typescript');
+const tsMorph = require('ts-morph');
+const cleanConsole = require('./mods/typer.js');
+
+const compilerOptions = {
+  allowJs: true,
+  module: ts.ModuleKind.ESNext,
+  target: ts.ScriptTarget.Latest,
+  newLine: ts.NewLineKind.LineFeed,
+  declaration: false,
+  noEmitOnError: false,
+  outDir: 'dist', // we're in-memory, but this is still required for some reason
+};
 
 /**
  * description
  *
  * @param {File} fileInfo - the file to run the rewriter on
- * @param {any} api - the API containing jscodeshift
  *
  * @return {String} the source of the newly rewritten file
  */
-function rewriter(fileInfo, api) {
-  const j = api.jscodeshift;
-  const root = j(fileInfo.source);
+function rewriter(fileInfo) {
+  const project = new tsMorph.Project({compilerOptions});
+  const sourceFile = project.createSourceFile(
+      fileInfo.path, fileInfo.source, {overwrite: true}
+  );
 
-//  cleanConsole(j, root);
-  const newSource = root.toSource({lineTerminator: '\n'});
+  sourceFile.forEachDescendant(cleanConsole);
 
-  const thirdSource = typer({path: fileInfo.path, source: newSource});
-
-  return thirdSource;
+  const result = project.emitToMemory();
+  const files = result.getFiles();
+  assert.strictEqual(files.length, 1);
+  return files[0].text;
 };
 
 module.exports = rewriter;
